Random selection of all heroes works w/ custom IDs
**Overview**

Using a custom `randotizerId` value for each hero that
gets generated on each app run. The values for this ID
are contiguous across all heros (unlike Valve's IDs).

This allows us to select a number in the range of IDs
we generate for our random hero.

**Also completed**

Some changes in HeroThumb.js and Home.js to use the
`activeHeroId` instead of the `isActive` flag to set
whether a hero is active.

**Not working**

Filtering based on int, str, and agi works as well as
randomly choosing a hero, but the two don't work together.

**Next steps**

Move the `randomHero()` function to another file and
expand its capabilities to support filtering of different
criteria, such as primary attribute. Note: this means the
`randomHero()` function will need to do more than just
spit out a number. It will need to be aware of all heros in
the list in order to filter and then randomly choose.